Re-add option to select decomposition type
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/abstraction-with-block-summaries@43558 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/algorithm/distributed_summaries/decomposition/MergeBlockNodesDecomposition.java
@@ -28,12 +28,12 @@ import org.sosy_lab.cpachecker.core.algorithm.distributed_summaries.decompositio
 public class MergeBlockNodesDecomposition implements CFADecomposer {
 
   private final Predicate<CFANode> isBlockEnd;
-  private final int targetNumber;
+  private final long targetNumber;
   private int id;
 
   private record BlockScope(CFANode start, CFANode last) {}
 
-  public MergeBlockNodesDecomposition(Predicate<CFANode> pIsBlockEnd, int pTargetNumber) {
+  public MergeBlockNodesDecomposition(Predicate<CFANode> pIsBlockEnd, long pTargetNumber) {
     isBlockEnd = pIsBlockEnd;
     targetNumber = pTargetNumber;
   }
